Some fix for terraform dependencies management

+++ iac/lab-account/us-east-1/prod/kubernetes-cluster/render.sh
@@ -5,20 +5,33 @@ set -x
 cd "$(dirname "$0")"
 
 
+
 TF_OUTPUT="$(yes | terragrunt output-all -json | jq -s add )"
-CLUSTER_NAME="$(echo ${TF_OUTPUT} | jq -r .cluster_name.value)"
-STATE="s3://$(echo ${TF_OUTPUT} | jq -r .kops_s3_bucket_name.value)"
+CLUSTER_NAME="$2"
+STATE="s3://$3"
 
-# Removing old etcd backups
-aws s3 rm ${STATE}/backups 
+yell() { echo "$0: $*" >&2; }
 
-kops toolbox template --name $CLUSTER_NAME --values <( echo ${TF_OUTPUT}) --template cluster-definition.yaml --format-yaml > cluster.yaml
+#Check if the output is a valid json
+if jq -e . >/dev/null 2>&1 <<<"$json_string"; then
+  # Removing old etcd backups
+  aws s3 rm ${STATE}/backups
 
-kops replace -f cluster.yaml --state ${STATE} --name ${CLUSTER_NAME} --force
+  kops toolbox template --name $CLUSTER_NAME \
+  --fail-on-missing=false \
+  --values <( echo ${TF_OUTPUT}) \
+  --set-string cluster_name=${CLUSTER_NAME} \
+  --set-string kops_s3_bucket_name=${STATE} \
+  --template cluster-definition.yaml --format-yaml > cluster.yaml
 
-kops create secret --name ${CLUSTER_NAME} --state ${STATE} sshpublickey admin -i ~/.ssh/id_rsa.pub
+  kops replace -f cluster.yaml --state ${STATE} --name ${CLUSTER_NAME} --force
 
-kops update cluster --out=. --target=terraform --state ${STATE} --name ${CLUSTER_NAME}
+  kops create secret --name ${CLUSTER_NAME} --state ${STATE} sshpublickey admin -i ~/.ssh/id_rsa.pub
 
-rm -fr versions.tf || true
-echo "yes" | terraform 0.12upgrade .
+  kops update cluster --out=. --target=terraform --state ${STATE} --name ${CLUSTER_NAME}
+
+  rm -fr versions.tf || true
+  echo "yes" | terraform 0.12upgrade .
+else
+  exit 0
+fi
